refactor: int, float -> Long, Float :: nullable

+++ back/tripDuo/src/main/java/com/example/tripDuo/entity/User.java
@@ -25,14 +25,14 @@ public class User {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     
     private String username;
     private String password;
     
     private String nickname;
               
-    private int age;
+    private Long age;
     private String name;
     private String gender;
     private String phoneNumber;
@@ -48,7 +48,7 @@ public class User {
               
     private String role; // [note: "user / manager / admin"]
 
-    private float ratings; // 지표 설정 
+    private Float ratings; // 지표 설정 
 
     private String lastLogin; // 몇분전 접속
